requirements.txt
google-generativeai>=0.7.2
python-dotenv
google-auth-oauthlib>=1.2.1
google-api-python-client
